python_migration_b/requirements.txt
azure-identity
azure-mgmt-apimanagement
azure-storage-blob
requests
python-dotenv
